postrepository수정, 중복검사 삭제

+++ src/main/java/com/tave8/ottu/dto/PostDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class PostDTO {
     private Long postIdx; // 수정에만 사용
     private Long userIdx;
-    private Long platformIdx;
+    private int platformIdx;
     private String content;
 
 }
